requirements.txt
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.1
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Pillow==6.2.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2018.7
uWSGI==2.0.18
whitenoise==4.1.4
